Import now working
I have now applied the logic that will allow you to import the CSV that
has been uploaded. In doing this though, I have discovered an issue
with the row_id not being pulled in to the system. Although not fatal,
this could cause an issue and I need to resolve.