Speeding up import
Initial method of importing was too slow. I estimated that it would
take around 18 hours to do 65,000 records. This was far too slow so I
have made changes to use arrays and import in bulk. I have also
performed the validation check before the import so that it does not
slow down when connecting to the model on the db.